+++ src/resources/AvailablePhoneNumbersBlocks.ts
@@ -1,0 +1,6 @@
+import TelnyxResource from '../TelnyxResource';
+
+export const AvailablePhoneNumbersBlocks = TelnyxResource.extend({
+  path: 'available_phone_number_blocks',
+  includeBasic: ['list'],
+});