+++ src/resources/PhoneNumbersRegulatoryRequirements.ts
@@ -1,7 +1,6 @@
 import TelnyxResource from '../TelnyxResource';
 
-export const PhoneNumberRegulatoryRequirements = TelnyxResource.extend({
+export const PhoneNumbersRegulatoryRequirements = TelnyxResource.extend({
   path: 'phone_numbers_regulatory_requirements',
-
   includeBasic: ['list'],
 });