+++ src/resources/PhoneNumbersSlim.ts
@@ -1,0 +1,6 @@
+import TelnyxResource from '../TelnyxResource';
+
+export const PhoneNumbersSlim = TelnyxResource.extend({
+  path: 'phone_numbers/slim',
+  includeBasic: ['list'],
+});