+++ src/test/resources/PhoneNumbers.test.ts
@@ -1,9 +1,9 @@
+import {TelnyxObject} from '../../Types';
 import {
   type ResponsePayloadList,
   type ResponsePayload,
   utils as testUtils,
 } from '../utils';
-const telnyx = testUtils.getTelnyxMock();
 
 const TEST_AUTH_KEY = testUtils.getUserTelnyxKey();
 const TEST_UUID = '123e4567-e89b-12d3-a456-426614174000';
@@ -15,6 +15,13 @@ type ResponsePayloadPhoneNumbers = ResponsePayload & {
 };
 
 describe('PhoneNumbers Resource', function () {
+  let telnyxInstance: TelnyxObject;
+
+  beforeEach(() => {
+    // make specs independent
+    telnyxInstance = testUtils.getTelnyxMock();
+  });
+
   function responseFn(response: ResponsePayload) {
     expect(response.data).toHaveProperty('id');
     expect(response.data).toHaveProperty('phone_number');
@@ -24,7 +31,7 @@ describe('PhoneNumbers Resource', function () {
   describe('update', function () {
     test('Sends the correct request', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
-      return telnyx.phoneNumbers
+      return telnyxInstance.phoneNumbers
         .update(TEST_UUID, {status: 'active'})
         .then(responseFn);
     });
@@ -33,7 +40,7 @@ describe('PhoneNumbers Resource', function () {
   describe('del', function () {
     test('Sends the correct request', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
-      return telnyx.phoneNumbers.del(TEST_UUID).then(responseFn);
+      return telnyxInstance.phoneNumbers.del(TEST_UUID).then(responseFn);
     });
   });
 
@@ -46,30 +53,32 @@ describe('PhoneNumbers Resource', function () {
 
     test('Sends the correct request', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
-      return telnyx.phoneNumbers.list().then(listResponseFn);
+      return telnyxInstance.phoneNumbers.list().then(listResponseFn);
     });
 
     test('Sends the correct request [with specified auth]', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
-      return telnyx.phoneNumbers.list(TEST_AUTH_KEY).then(listResponseFn);
+      return telnyxInstance.phoneNumbers
+        .list(TEST_AUTH_KEY)
+        .then(listResponseFn);
     });
   });
 
   describe('retrieve', function () {
     test('Sends the correct request', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
-      return telnyx.phoneNumbers.retrieve(TEST_UUID).then(responseFn);
+      return telnyxInstance.phoneNumbers.retrieve(TEST_UUID).then(responseFn);
     });
 
     test('Sends the correct request [with specified auth]', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
-      return telnyx.phoneNumbers
+      return telnyxInstance.phoneNumbers
         .retrieve(TEST_UUID, TEST_AUTH_KEY)
         .then(responseFn);
     });
   });
 
-  describe('setEmergencySettings', function () {
+  describe('enableEmergencySettings', function () {
     function responseFn(response: ResponsePayloadPhoneNumbers) {
       expect(response.data).toHaveProperty('id');
       expect(response.data).toMatchObject({record_type: 'voice_settings'});
@@ -82,8 +91,8 @@ describe('PhoneNumbers Resource', function () {
 
     test('Sends the correct request', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
-      return telnyx.phoneNumbers
-        .setEmergencySettings(TEST_UUID, {
+      return telnyxInstance.phoneNumbers
+        .enableEmergencySettings(TEST_UUID, {
           emergency_enabled: true,
           emergency_address_id: '1315261609962112019',
         })
@@ -92,8 +101,8 @@ describe('PhoneNumbers Resource', function () {
 
     test('Sends the correct request [with specified auth]', function () {
       // @ts-expect-error TODO: import .d.ts files under src/test folder
-      return telnyx.phoneNumbers
-        .setEmergencySettings(
+      return telnyxInstance.phoneNumbers
+        .enableEmergencySettings(
           TEST_UUID,
           {
             emergency_enabled: true,
@@ -116,14 +125,14 @@ describe('PhoneNumbers Resource', function () {
     describe('retrieveVoiceSettings', function () {
       test('Sends the correct request', function () {
         // @ts-expect-error TODO: import .d.ts files under src/test folder
-        return telnyx.phoneNumbers
+        return telnyxInstance.phoneNumbers
           .retrieveVoiceSettings(TEST_UUID)
           .then(responseFn);
       });
 
       test('Sends the correct request [with specified auth]', function () {
         // @ts-expect-error TODO: import .d.ts files under src/test folder
-        return telnyx.phoneNumbers
+        return telnyxInstance.phoneNumbers
           .retrieveVoiceSettings(TEST_UUID, TEST_AUTH_KEY)
           .then(responseFn);
       });
@@ -132,7 +141,7 @@ describe('PhoneNumbers Resource', function () {
     describe('updateVoiceSettings', function () {
       test('Sends the correct request', function () {
         // @ts-expect-error TODO: import .d.ts files under src/test folder
-        return telnyx.phoneNumbers
+        return telnyxInstance.phoneNumbers
           .updateVoiceSettings(TEST_UUID, {
             tech_prefix_enabled: true,
           })
@@ -152,14 +161,14 @@ describe('PhoneNumbers Resource', function () {
     describe('retrieveMessagingSettings', function () {
       test('Sends the correct request', function () {
         // @ts-expect-error TODO: import .d.ts files under src/test folder
-        return telnyx.phoneNumbers
+        return telnyxInstance.phoneNumbers
           .retrieveMessagingSettings(TEST_UUID)
           .then(responseFn);
       });
 
       test('Sends the correct request [with specified auth]', function () {
         // @ts-expect-error TODO: import .d.ts files under src/test folder
-        return telnyx.phoneNumbers
+        return telnyxInstance.phoneNumbers
           .retrieveMessagingSettings(TEST_UUID, TEST_AUTH_KEY)
           .then(responseFn);
       });
@@ -168,9 +177,55 @@ describe('PhoneNumbers Resource', function () {
     describe('updateMessagingSettings', function () {
       test('Sends the correct request', function () {
         // @ts-expect-error TODO: import .d.ts files under src/test folder
-        return telnyx.phoneNumbers
+        return telnyxInstance.phoneNumbers
           .updateMessagingSettings(TEST_UUID, {
             messaging_product: 'P2P',
+          })
+          .then(responseFn);
+      });
+    });
+  });
+
+  describe('Voicemail methods', function () {
+    function responseFn(response: ResponsePayload) {
+      expect(response.data).toHaveProperty('enabled');
+      expect(response.data).toHaveProperty('pin');
+    }
+
+    describe('retrieveVoicemail', function () {
+      test('Sends the correct request', function () {
+        // @ts-expect-error TODO: import .d.ts files under src/test folder
+        return telnyxInstance.phoneNumbers
+          .retrieveVoicemail(TEST_UUID)
+          .then(responseFn);
+      });
+
+      test('Sends the correct request [with specified auth]', function () {
+        // @ts-expect-error TODO: import .d.ts files under src/test folder
+        return telnyxInstance.phoneNumbers
+          .retrieveVoicemail(TEST_UUID, TEST_AUTH_KEY)
+          .then(responseFn);
+      });
+    });
+
+    describe('createVoicemail', function () {
+      test('Sends the correct request', function () {
+        // @ts-expect-error TODO: import .d.ts files under src/test folder
+        return telnyxInstance.phoneNumbers
+          .createVoicemail(TEST_UUID, {
+            pin: '1234',
+            enabled: true,
+          })
+          .then(responseFn);
+      });
+    });
+
+    describe('updateVoicemail', function () {
+      test('Sends the correct request', function () {
+        // @ts-expect-error TODO: import .d.ts files under src/test folder
+        return telnyxInstance.phoneNumbers
+          .updateVoicemail(TEST_UUID, {
+            pin: '1234',
           })
           .then(responseFn);
       });